+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - Support Fan light temperature and color. (#184)
 - Support Humidifier light. (#184)
 - Expose energy usage for outlets/switches. (#190) Thanks @lstrojny for the contribution
+- Strict config validate for `deviceOverrides`.
 
 
 ## [1.6.0] - (2022.12.3)